packages/py/pyproject.toml
[tool.poetry]
name = "taskless_shim"
version = "0.1.0"
description = "A python shim for L7 observability"
authors = ["Taskless <oss@taskless.io>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
requests = ">=2.30.0"
urlmatch = "^1.0.1"
responses = "^0.25.3"
snowflake-id = "^1.0.2"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
